cfg-lexer: extract CfgBlockGenerator into a module of its own
