+++ lib/Makefile.am
@@ -97,6 +97,8 @@ pkginclude_HEADERS			+= \
 	lib/cfg-lexer.h			\
 	lib/cfg-lexer-subst.h		\
 	lib/cfg-args.h			\
+	lib/cfg-block.h			\
+	lib/cfg-block-generator.h	\
 	lib/cfg-parser.h		\
 	lib/cfg-tree.h			\
 	lib/children.h			\
@@ -178,6 +180,8 @@ lib_libsyslog_ng_la_SOURCES		= \
 	lib/cache.c			\
 	lib/cfg.c			\
 	lib/cfg-args.c			\
+	lib/cfg-block.c			\
+	lib/cfg-block-generator.c	\
 	lib/cfg-lexer.c			\
 	lib/cfg-lexer-subst.c		\
 	lib/cfg-parser.c		\
