+++ modules/confgen/confgen-plugin.c
@@ -24,6 +24,7 @@
 #include "confgen.h"
 #include "cfg.h"
 #include "cfg-lexer.h"
+#include "cfg-block-generator.h"
 #include "messages.h"
 #include "plugin.h"
